Updating features to work with mpas-lettuce super-repo layout
These are changes to migrate from the old setup of lettuce to the newer version
that moves a lot of common tasks into a framework level.  Changes include
renaming scenario language to be consistent with new conventions and rewriting
most of the get_test_case function in landice_tasks.py following the version
in ocean_tasks.py as a template.
The Halfar feature does not work correctly yet.